requirements.txt
alembic==0.9.1
apipkg==1.4
appdirs==1.4.3
Babel==2.4.0
blinker==1.4
click==6.7
colorama==0.3.7
execnet==1.4.1
Flask==0.12.1
Flask-Babel==0.11.1
Flask-Compress==1.4.0
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==2.0.3
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.5
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
gunicorn==19.7.1
hurry.filesize==0.9
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
Mako==1.0.6
MarkupSafe==1.0
nose==1.3.7
packaging==16.8
passlib==1.7.1
py==1.4.33
py-bcrypt==0.4
pyparsing==2.2.0
pytest==3.0.7
pytest-flask==0.10.0
pytest-xdist==1.15.0
python-editor==1.0.3
python-termstyle==0.1.10
pytz==2017.2
six==1.10.0
sniffer==0.4.0
speaklater==1.3
SQLAlchemy==1.1.8
Unidecode==0.4.20
-e git+git@bitbucket.org:voxolab/voxo-lib.git@master#egg=voxo_lib
Werkzeug==0.12.1
WTForms==2.1
